Use ptrdiff_t for casts to pointers.
Some casts to a pointer type were first casting to "long". This is
a good way to prevent a warning about casting to a different size -
but only on platforms where "long" really has the same size as a
pointer. On MingGW64 a long has only 4 bytes. Sometimes a "long" is
just not long enough.