Keep sequence order of input tables in TableFlatten.
TableFlatten iterates over the elements in an input table with
lua_next. There is no guarantee that the elements are processed
in an ordered way, also not for sequences. There is a rare chance
that TableFlatten mixes up the order of a sequence.

With this patch TableFlatten loops only over the sequence part of a
table. This keeps the order of the input sequences, but also drops
all other key-value pairs.